# openapi/cult-api.yaml
openapi: 3.1.0
info:
  title: CULT API
  version: 0.1.0
  description: Private backend API for CULT MVP 1.
servers:
  - url: http://localhost:3001
paths:
  /health:
    get:
      summary: Liveness check
      responses:
        "200":
          description: Service is alive
  /ready:
    get:
      summary: Readiness check
      responses:
        "200":
          description: Service is ready
        "503":
          description: Service is not ready
  /v1/events:
    get:
      summary: Discover cultural events
      description: >-
        M7 Discovery API. Every parameter below is implemented — none return
        `filter-not-implemented`. Defaults to `status=scheduled`
        (cancelled/postponed/rescheduled/completed events are excluded unless `status` is
        given explicitly). `period` and `start`/`end` are mutually exclusive; `start` and
        `end` must be given together. `lat`/`lng` must be given together; `radius` requires
        both. All calendar-date semantics (`period`, `start`, `end`) resolve in
        America/Sao_Paulo, never UTC or the server's host timezone.
      parameters:
        - in: query
          name: q
          description: Full-text-ish search across title, description, venue name and performer names.
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: period
          description: >-
            A named discovery shortcut, resolved against the current date in
            America/Sao_Paulo. Cannot be combined with `start`/`end`.
          schema:
            type: string
            enum: [today, tomorrow, weekend, this_week, this_month]
        - in: query
          name: start
          description: Inclusive start of a custom date range (YYYY-MM-DD). Must be given with `end`.
          schema:
            type: string
            format: date
        - in: query
          name: end
          description: Inclusive end of a custom date range (YYYY-MM-DD). Must be given with `start`.
          schema:
            type: string
            format: date
        - in: query
          name: free
          schema:
            type: boolean
        - in: query
          name: lat
          description: Must be given with `lng`.
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - in: query
          name: lng
          description: Must be given with `lat`.
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - in: query
          name: radius
          description: Search radius in meters. Requires `lat`/`lng`. Defaults to 5000.
          schema:
            type: integer
            minimum: 1
            maximum: 50000
            default: 5000
        - in: query
          name: status
          description: Defaults to `scheduled` when omitted.
          schema:
            $ref: "#/components/schemas/EventStatus"
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Event page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /v1/events/{slug}:
    get:
      summary: Get one cultural event by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/analytics:
    post:
      summary: Record a first-party product analytics event
      description: >-
        M10 sections 25-28. First-party, minimal, best-effort product telemetry — never
        precise geolocation, never anything that could re-identify a specific
        request/device. Called by apps/web's own `/api/analytics` BFF route server-side, not
        directly by the browser. `metadata` accepts only an explicit allowlisted set of keys
        (see AnalyticsMetadata below); an unlisted key is rejected (400), never silently
        dropped. Not an operational/administrative endpoint like M9's banned
        `POST /v1/dedup/...` — this never mutates product/discovery state, only appends a
        telemetry row.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyticsEventInput"
      responses:
        "202":
          description: Event accepted
        "400":
          $ref: "#/components/responses/Problem"
        "413":
          description: Payload too large
        "500":
          $ref: "#/components/responses/Problem"

  /v1/categories:
    get:
      summary: List event categories
      description: >-
        Distinct category ids actually present on existing events — never a fixed/invented
        taxonomy. There is no categories table or human-readable name source yet, so `name`
        and `slug` currently mirror the raw technical `id` (see Category below).
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"

components:
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"

  schemas:
    EventStatus:
      type: string
      enum:
        - scheduled
        - cancelled
        - postponed
        - rescheduled
        - completed

    EventListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        pagination:
          type: object
          properties:
            next_cursor:
              type: [string, "null"]

    Event:
      type: object
      required:
        - id
        - slug
        - title
        - status
        - occurrences
        - sources
        - quality_score
        - ranking_score
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/EventStatus"
        category:
          type: [string, "null"]
        occurrences:
          type: array
          items:
            $ref: "#/components/schemas/EventOccurrence"
        venue:
          anyOf:
            - $ref: "#/components/schemas/Venue"
            - type: "null"
        free:
          type: [boolean, "null"]
        price_min:
          type: [number, "null"]
        price_max:
          type: [number, "null"]
        currency:
          type: [string, "null"]
          example: BRL
        image_url:
          type: [string, "null"]
          format: uri
        ticket_url:
          type: [string, "null"]
          format: uri
        sources:
          type: array
          items:
            $ref: "#/components/schemas/EventSource"
        quality_score:
          type: number
          minimum: 0
          maximum: 1
        ranking_score:
          type: number
        distance_meters:
          description: Only present when the request included lat/lng/radius (nearby search).
          type: number

    # M4 / ADR-0014: not every source gives a precise instant (e.g. a multi-day exhibition
    # with no published time). "timed" carries a real date-time; "date" carries calendar
    # dates only — never a time of "00:00" invented to fit the old shape.
    EventOccurrence:
      oneOf:
        - $ref: "#/components/schemas/TimedEventOccurrence"
        - $ref: "#/components/schemas/DateOnlyEventOccurrence"
      discriminator:
        propertyName: kind
        mapping:
          timed: "#/components/schemas/TimedEventOccurrence"
          date: "#/components/schemas/DateOnlyEventOccurrence"

    TimedEventOccurrence:
      type: object
      required: [kind, starts_at, timezone, status]
      properties:
        kind:
          type: string
          enum: [timed]
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: [string, "null"]
          format: date-time
        timezone:
          type: string
          example: America/Sao_Paulo
        status:
          $ref: "#/components/schemas/EventStatus"

    DateOnlyEventOccurrence:
      type: object
      required: [kind, start_date, timezone, status]
      properties:
        kind:
          type: string
          enum: [date]
        start_date:
          type: string
          format: date
          example: "2026-08-29"
        end_date:
          type: [string, "null"]
          format: date
          example: "2026-09-20"
        timezone:
          type: string
          example: America/Sao_Paulo
        status:
          $ref: "#/components/schemas/EventStatus"

    Venue:
      type: object
      required: [name, city, state, country]
      properties:
        name:
          type: string
        address:
          type: [string, "null"]
        neighborhood:
          type: [string, "null"]
        city:
          type: string
        state:
          type: string
        country:
          type: string
        latitude:
          type: [number, "null"]
        longitude:
          type: [number, "null"]

    EventSource:
      type: object
      required: [source_id, url, confidence]
      properties:
        source_id:
          type: string
        external_id:
          type: [string, "null"]
        url:
          type: string
          format: uri
        confidence:
          type: number
          minimum: 0
          maximum: 1

    Category:
      type: object
      description: >-
        `name` and `slug` currently mirror `id` — no human-readable taxonomy exists yet
        (MVP1: category ids are free-text, provider-derived; see CLAUDE.md).
      required: [id, name, slug]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string

    AnalyticsEventInput:
      type: object
      required: [event_name]
      properties:
        event_name:
          type: string
          enum:
            - page_view
            - event_view
            - search
            - filter_used
            - nearby_used
            - map_opened
            - share
            - ticket_click
            - maps_click
        event_id:
          type: string
          description: Optional — the CanonicalEvent this event relates to, when applicable.
        metadata:
          type: object
          description: >-
            Only allowlisted keys are accepted: period, category, free. Any other key is
            rejected with 400, never silently dropped. Never latitude/longitude or any other
            precise/re-identifying value.
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: boolean

    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string